+++ test/functional/android/android/image_comparison_test.rb
@@ -100,8 +100,7 @@ class AppiumLibCoreTest
 
         assert image_element.inspect
         assert image_element.hash
-        _, element_id = image_element.ref
-        assert element_id =~ /\Aappium-image-element-[a-z0-9\-]+/
+        assert image_element.id =~ /\Aappium-image-element-[a-z0-9\-]+/
 
         el_location = el.location
         image_location = image_element.location
@@ -144,8 +143,7 @@ class AppiumLibCoreTest
 
         assert image_element.inspect
         assert image_element.hash
-        _, element_id = image_element.ref
-        assert element_id =~ /\Aappium-image-element-[a-z0-9\-]+/
+        assert image_element.id =~ /\Aappium-image-element-[a-z0-9\-]+/
 
         el_location = el.location
         image_location = image_element.location
@@ -187,8 +185,7 @@ class AppiumLibCoreTest
         image_element = @driver.find_element_by_image AppiumLibCoreTest.path_of('test/functional/data/test_android_nfc_270.png')
         assert image_element.inspect
         assert image_element.hash
-        _, element_id = image_element.ref
-        assert element_id =~ /\Aappium-image-element-[a-z0-9\-]+/
+        assert image_element.id =~ /\Aappium-image-element-[a-z0-9\-]+/
 
         el_location = el.location
         image_location = image_element.location