+++ test/functional/android/android/mobile_commands_test.rb
@@ -97,11 +97,9 @@ class AppiumLibCoreTest
         @driver = @core.start_driver
 
         el = @core.wait { @driver.find_element(:accessibility_id, 'Views') }
-        _, element_id = el.ref
-
-        assert_mobile_command_error 'mobile: openDrawer', { element: element_id, gravity: 1 },
+        assert_mobile_command_error 'mobile: openDrawer', { element: el.id, gravity: 1 },
                                     'Could not open drawer'
-        assert_mobile_command_error 'mobile: closeDrawer', { element: element_id, gravity: 1 },
+        assert_mobile_command_error 'mobile: closeDrawer', { element: el.id, gravity: 1 },
                                     'Could not close drawer'
       end
 
@@ -117,8 +115,7 @@ class AppiumLibCoreTest
         @core.wait { @driver.find_element(:accessibility_id, 'change the date') }.click
 
         date_picker = @driver.find_element(:id, 'android:id/datePicker')
-        _, element_id = date_picker.ref
-        @driver.execute_script('mobile: setDate', { year: 2020, monthOfYear: 10, dayOfMonth: 25, element: element_id })
+        @driver.execute_script('mobile: setDate', { year: 2020, monthOfYear: 10, dayOfMonth: 25, element: date_picker.id })
         assert_equal 'Sun, Oct 25', @driver.find_element(:id, 'android:id/date_picker_header_date').text
       end
 
@@ -132,13 +129,11 @@ class AppiumLibCoreTest
         @driver = @core.start_driver
 
         time_el = @core.wait { @driver.find_element(:class, 'android.widget.TimePicker') }
-        _, element_id = time_el.ref
-        @driver.execute_script('mobile: setTime', { hours: 11, minutes: 0, element: element_id })
+        @driver.execute_script('mobile: setTime', { hours: 11, minutes: 0, element: time_el.id })
         assert @driver.find_element(:id, 'io.appium.android.apis:id/dateDisplay').text == '11:00'
 
         time_el = @driver.find_element(:class, 'android.widget.TimePicker')
-        _, element_id = time_el.ref
-        @driver.execute_script('mobile: setTime', { hours: 15, minutes: 15, element: element_id })
+        @driver.execute_script('mobile: setTime', { hours: 15, minutes: 15, element: time_el.id })
         assert @driver.find_element(:id, 'io.appium.android.apis:id/dateDisplay').text == '15:15'
       end
 
@@ -151,13 +146,12 @@ class AppiumLibCoreTest
 
         el = @core.wait { @driver.find_element(:accessibility_id, 'Views') }
 
-        _, element_id = el.ref
-        assert_mobile_command_error 'mobile: navigateTo', { element: element_id, menuItemId: -100 },
+        assert_mobile_command_error 'mobile: navigateTo', { element: el.id, menuItemId: -100 },
                                     'must be a non-negative number'
-        assert_mobile_command_error 'mobile: navigateTo', { element: element_id, menuItemId: 'no element' },
+        assert_mobile_command_error 'mobile: navigateTo', { element: el.id, menuItemId: 'no element' },
                                     'must be a non-negative number'
         # A test demo apk has no the element
-        assert_mobile_command_error 'mobile: navigateTo', { element: element_id, menuItemId: 10 },
+        assert_mobile_command_error 'mobile: navigateTo', { element: el.id, menuItemId: 10 },
                                     'Could not navigate to menu item 10'
       end
 
@@ -170,20 +164,19 @@ class AppiumLibCoreTest
         @driver = @core.start_driver
         el = @core.wait { @driver.find_element(:accessibility_id, 'Views') }
 
-        _, element_id = el.ref
-        assert_mobile_command_error 'mobile: scrollToPage',  { element: element_id, scrollTo: 'right' },
+        assert_mobile_command_error 'mobile: scrollToPage',  { element: el.id, scrollTo: 'right' },
                                     'Could not perform scroll to on element'
-        assert_mobile_command_error 'mobile: scrollToPage',  { element: element_id, scrollTo: '' },
+        assert_mobile_command_error 'mobile: scrollToPage',  { element: el.id, scrollTo: '' },
                                     'Invalid scrollTo parameters'
-        assert_mobile_command_error 'mobile: scrollToPage',  { element: element_id, scrollToPage: -100 },
+        assert_mobile_command_error 'mobile: scrollToPage',  { element: el.id, scrollToPage: -100 },
                                     'be a non-negative integer'
         error = assert_raises ::Selenium::WebDriver::Error::InvalidArgumentError do
-          @driver.execute_script 'mobile: scrollToPage', { element: element_id }
+          @driver.execute_script 'mobile: scrollToPage', { element: el.id }
         end
         assert error.message.include? "Must provide either 'scrollTo' or 'scrollToPage'"
 
         # A test demo apk has no the element
-        assert_mobile_command_error 'mobile: scrollToPage', { element: element_id, scrollToPage: 2, smoothScroll: true },
+        assert_mobile_command_error 'mobile: scrollToPage', { element: el.id, scrollToPage: 2, smoothScroll: true },
                                     'Could not perform scroll to on element'
       end
 
@@ -203,9 +196,8 @@ class AppiumLibCoreTest
         e = @driver.find_elements :class, 'android.widget.TextView'
         assert_equal '0', e.last.text
 
-        _, element_id = e.last.ref
         type = @driver.execute_script('mobile: backdoor',
-                                      { target: :element, elementId: element_id, methods: [{ name: 'getTypeface' }] })
+                                      { target: :element, elementId: e.last.id, methods: [{ name: 'getTypeface' }] })
         assert type['mStyle']
       end
 
@@ -219,10 +211,9 @@ class AppiumLibCoreTest
         @driver = @core.start_driver
 
         el = @core.wait { @driver.find_element(:id, 'wv1') }
-        _, element_id = el.ref
 
         @driver.execute_script 'mobile: webAtoms', {
-          webviewElement: element_id,
+          webviewElement: el.id,
           forceJavascriptEnabled: true,
           methodChain: [
             { name: 'withElement', atom: { name: 'findElement', locator: { using: 'ID', value: 'i_am_a_textbox' } } },
@@ -235,7 +226,7 @@ class AppiumLibCoreTest
         # Can use `atom` defined in https://developer.android.com/reference/android/support/test/espresso/web/webdriver/DriverAtoms
         # Locator: https://developer.android.com/reference/androidx/test/espresso/web/webdriver/Locator
         @driver.execute_script 'mobile: webAtoms', {
-          webviewElement: element_id,
+          webviewElement: el.id,
           forceJavascriptEnabled: true,
           methodChain: [{
             name: 'withElement',
@@ -246,7 +237,7 @@ class AppiumLibCoreTest
         # Raises an error if the method cannot find any DriverAtoms or necessary arguments
         error = assert_raises ::Selenium::WebDriver::Error::WebDriverError do
           @driver.execute_script 'mobile: webAtoms', {
-            webviewElement: element_id,
+            webviewElement: el.id,
             forceJavascriptEnabled: true,
             methodChain: [{
               name: 'withElement',