+++ test/functional/ios/ios/device_test.rb
@@ -206,8 +206,7 @@ class AppiumLibCoreTest
         if @@core.automation_name == :xcuitest
           element = @@core.wait { @@driver.find_element :accessibility_id, ACTIVITY_INDICATORS }
 
-          _, element_id = element.ref
-          @@driver.execute_script('mobile: tap', x: 0, y: 0, element: element_id)
+          @@driver.execute_script('mobile: tap', x: 0, y: 0, element: element.id)
         else
           Appium::Core::TouchAction.new(@@driver)
                                    .press(element: @@driver.find_element(:accessibility_id, ACTIVITY_INDICATORS))